requirements.txt
pybit
python-dotenv
requests
